+++ corehq/apps/users/decorators.py
@@ -2,9 +2,9 @@ from django.http import Http404, HttpResponse
 from django.core.exceptions import PermissionDenied
 from corehq.apps.domain.decorators import login_and_domain_required, domain_specific_login_redirect
 from functools import wraps
-from corehq.apps.users.models import CouchUser
+from corehq.apps.users.models import CouchUser, CommCareUser
 from django.utils.translation import ugettext as _
-
+from dimagi.utils.web import json_response
 
 def require_permission_raw(permission_check, login_decorator=login_and_domain_required):
     """
@@ -79,3 +79,36 @@ def require_permission_to_edit_user(view_func):
         else:
             return domain_specific_login_redirect(request, domain)
     return _inner
+
+def ensure_active_user():
+    def decorator(fn):
+        @wraps(fn)
+        def _inner(request, *args, **kwargs):
+            """
+            ensures active user for sync and form receive requests
+            returns the error code which should be present in app_string for translation and also shares
+            the default response which can be used in case the translation is missing
+            """
+            couch_user = CouchUser.from_django_user(request.user)
+            if couch_user:
+                valid, message, error_code = ensure_active_user_by_username(couch_user.username)
+
+            if not valid:
+                # respond with apt error message and status code 406(unacceptable)
+                return json_response({
+                    "error": error_code,
+                    "default_response": message
+                }, status_code=406)
+            else:
+                return fn(request, *args, **kwargs)
+        return _inner
+    return decorator
+
+def ensure_active_user_by_username(username):
+    ccu = CommCareUser.get_by_username(username)
+    valid, message, error_code = True, None, None
+    if ccu and not ccu.is_active:
+        valid, message, error_code = False, 'User deactivated', 'mobile.app.translation.user.is.deactivated'
+    elif CommCareUser.get_deleted_by_username(username):
+        valid, message, error_code = False, 'User deleted', 'mobile.app.translation.user.is.deleted'
+    return valid, message, error_code
